move scale types to scale.ts
